// zeus-client-visual-programming/src/app/enums/query-keys.enum.ts
export enum QueryKeys {
  PROJECTS = 'projects/',
  PROJECT = 'project/',
  DESIGNER_WORKSPACE = 'designerWorkspace/',
  BLUEPRINT_COMPONENTS = 'blueprintComponents/',
  VISUALIZER_WORKSPACE = 'visualizerWorkspace/',
  VISUALIZER_WORKSPACE_CONNECTIONS = 'visualizerWorkspaceConnections/',
  PROJECT_CODE_MODULES = 'projectCodeModules/',
  SYSTEM_CODE_MODULES = 'systemCodeModules/',
  EXPORTED_PROJECTS = 'exportedProjects/'
}